Delete sub tasks and notes inside a project once deleted

+++ src/services/DeleteData.jsx
@@ -1,6 +1,6 @@
-import { doc, deleteDoc } from 'firebase/firestore';
+import { doc, deleteDoc, where, getDocs, collection, query } from 'firebase/firestore';
 import { db } from '../config/firebase';
-import { useReloadContext } from '../context/ReloadContext';
+
 
 const deleteData = async ({id, collectionName, reloadComponent}) => {
   try {
@@ -12,7 +12,39 @@ const deleteData = async ({id, collectionName, reloadComponent}) => {
     }
 
     const docRef = doc(db, collectionName, id);
-    await deleteDoc(docRef);
+    if(docRef) {
+      try {
+        await deleteDoc(docRef);
+      } catch(error) {
+        alert(error);
+        console.log(error);
+      }
+    } else {
+      console.log('Cannot access doc or it doesnt exist')
+    }
+
+    const subTaskRef = collection(db, 'tasks');
+    const subTaskQ = query(subTaskRef, where('project-id', '==', id));
+    const subTaskSnap = await getDocs(subTaskQ);
+
+    const subNoteRef = collection(db, 'notes')
+    const subNotesQ = query(subNoteRef, where('project-id', '==', id));
+    const subNoteSnap = await getDocs(subNotesQ);
+
+    if(subNoteSnap, subTaskSnap) {
+      const taskDelete = subTaskSnap.docs.map(task => deleteDoc(task.ref));
+      const noteDelete = subNoteSnap.docs.map(notes => deleteData(note.ref));
+
+      try {
+        await Promise.all([...taskDelete, ...noteDelete])
+      } catch(error) {
+        alert(error);
+        console.log(error);
+      }
+    } else {
+      console.log('No sub tasks/notes or error accesing sub collections')
+    }
+
     reloadComponent();
     alert("Successfully Deleted");
